requirements.txt
pandas
scikit-learn
nltk
joblib
streamlit
newspaper3k
